+++ src/app/email/verify/[id]/page.tsx
@@ -1,34 +1,10 @@
-'use client'
+import { Metadata } from 'next'
+import { EmailVerify } from './EmailVerify'
 
-import { noAuthMiddleware } from '@/hocs/noAuthMiddleware'
-import axios from 'axios'
-import { useParams, useRouter, useSearchParams } from 'next/navigation'
-import { useAuthUserMutators } from '@/store/authUserState'
-
-const VerifyEmail = () => {
-  const router = useRouter()
-  const params = useParams()
-  const searchParams = useSearchParams()
-
-  const { updateAuthUser } = useAuthUserMutators()
-
-  ;(async () => {
-    try {
-      const response = await axios.get(
-        `/email/verify/${params.id}?expires=${searchParams.get(
-          'expires'
-        )}&signature=${searchParams.get('signature')}`
-      )
-
-      updateAuthUser(response.data.data)
-      // store.dispatch('notice')
-    } catch (error) {
-      console.log(error)
-    }
-    router.push('/')
-  })()
-
-  return <div>メール認証中...</div>
+export const metadata: Metadata = {
+  title: 'メール認証',
 }
 
-export default noAuthMiddleware(VerifyEmail)
+export default function Page() {
+  return <EmailVerify />
+}
