+++ src/app/login/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata } from 'next'
-import Login from './Login'
+import { Login } from './Login'
 
 export const metadata: Metadata = {
   title: 'ログイン',
