+++ src/app/pokemons/[id]/PokemonShow.tsx
@@ -1,0 +1,150 @@
+'use client'
+
+import { useEffect, useState } from 'react'
+import { useParams, useRouter } from 'next/navigation'
+import axios from 'axios'
+import {
+  Box,
+  Button,
+  Container,
+  Table,
+  TableBody,
+  TableCell,
+  TableRow,
+  Typography,
+} from '@mui/material'
+import Link from 'next/link'
+import { PokemonBasicInfo, PokemonSummary } from '@/types'
+import { LoadingPageTemplate } from '@/components/templates/LoadingPageTemplate'
+import { useAuthUserState } from '@/store/authUserState'
+import { usePokemonBasicInfosState } from '@/store/pokemonBasicInfosState'
+
+export const PokemonShow = () => {
+  const params = useParams()
+
+  const router = useRouter()
+  const authUser = useAuthUserState()
+  const pokemonBasicInfos = usePokemonBasicInfosState()
+
+  const [pokemonSummary, setPokemonSummary] = useState<PokemonSummary | null>(
+    null
+  )
+
+  const [pokemonBasicInfo, setPokemonBasicInfo] =
+    useState<PokemonBasicInfo | null>(null)
+
+  useEffect(() => {
+    ;(async () => {
+      try {
+        const response = await axios.get<{ data: PokemonSummary }>(
+          `/pokemons/${params.id}`
+        )
+        setPokemonSummary(response.data.data)
+      } catch (error) {
+        router.push('/')
+      }
+    })()
+  }, [params.id, router])
+
+  useEffect(() => {
+    const targetPokemonBasicInfo = pokemonBasicInfos.find(
+      (basicInfo) => basicInfo.name === pokemonSummary?.name
+    )
+
+    if (targetPokemonBasicInfo === undefined) return
+
+    setPokemonBasicInfo(targetPokemonBasicInfo)
+  }, [pokemonBasicInfos, pokemonSummary])
+
+  const handleEditItem = (item: PokemonSummary) => {
+    if (item.user.username === authUser?.username) {
+      router.push(`/pokemons/${item.id}/edit`)
+    } else {
+      router.push('/')
+    }
+  }
+
+  const handleDeleteItem = async (id: number) => {
+    try {
+      await axios.delete(`/pokemons/${id}`)
+      router.push(`/users/${authUser?.username}`)
+    } catch (error) {
+      router.push('/')
+    }
+  }
+
+  if (pokemonSummary === null || !pokemonBasicInfo)
+    return <LoadingPageTemplate />
+
+  return (
+    <Container>
+      <Typography>
+        ポケモンの詳細ページは、現在コンテンツの表示を調整中です。
+      </Typography>
+      <Table>
+        <TableBody>
+          <TableRow>
+            <TableCell align="left">No</TableCell>
+            <TableCell align="right">{pokemonBasicInfo.no}</TableCell>
+          </TableRow>
+          <TableRow>
+            <TableCell align="left">ポケモン名</TableCell>
+            <TableCell align="right">{pokemonSummary.name}</TableCell>
+          </TableRow>
+          <TableRow>
+            <TableCell align="left">レベル</TableCell>
+            <TableCell align="right">{pokemonSummary.lv}</TableCell>
+          </TableRow>
+          <TableRow>
+            <TableCell align="left">性格</TableCell>
+            <TableCell align="right">{pokemonSummary.nature}</TableCell>
+          </TableRow>
+          <TableRow>
+            <TableCell align="left">種族値</TableCell>
+            <TableCell align="right">
+              {pokemonBasicInfo.baseStats.hp}-
+              {pokemonBasicInfo.baseStats.attack}-
+              {pokemonBasicInfo.baseStats.defense}-
+              {pokemonBasicInfo.baseStats.spAttack}-
+              {pokemonBasicInfo.baseStats.spDefense}-
+              {pokemonBasicInfo.baseStats.speed}
+            </TableCell>
+          </TableRow>
+          <TableRow>
+            <TableCell align="left">ステータス</TableCell>
+            <TableCell align="right">{pokemonSummary.stats}</TableCell>
+          </TableRow>
+        </TableBody>
+      </Table>
+      {pokemonSummary.description && (
+        <>
+          <span>ポケモンの説明</span>
+          <p>{pokemonSummary.description}</p>
+        </>
+      )}
+      {pokemonSummary.user.username === authUser?.username ? (
+        <Box display="flex" justifyContent="center">
+          <Button
+            color="primary"
+            onClick={() => handleEditItem(pokemonSummary)}
+          >
+            編集
+          </Button>
+          <Button
+            color="error"
+            onClick={() => handleDeleteItem(pokemonSummary.id)}
+          >
+            削除
+          </Button>
+        </Box>
+      ) : (
+        <Typography>
+          投稿者：
+          <Link href={`/users/${pokemonSummary.user.username}`}>
+            {pokemonSummary.user.nickname}
+          </Link>
+        </Typography>
+      )}
+    </Container>
+  )
+}
