+++ src/app/pokemons/[id]/edit/PokemonEdit.tsx
@@ -1,0 +1,5 @@
+'use client'
+
+export const PokemonEdit = () => {
+  return <div>ポケモン編集</div>
+}
