+++ src/app/register/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata } from 'next'
-import Register from './Register'
+import { Register } from './Register'
 
 export const metadata: Metadata = {
   title: 'アカウント作成',
