+++ src/app/settings/Settings.tsx
@@ -1,0 +1,7 @@
+'use client'
+
+import { authMiddleware } from '@/hocs/authMiddleware'
+
+export const Settings = authMiddleware(() => {
+  return <>設定</>
+})
